refactor: add detection module
