+++ apps/extension/src/background.js
@@ -281,1508 +281,15 @@ async function checkAllPlatformsProgressive(platforms, tabId) {
   }
 }
 
+import { detectUser } from '@cose/detection'
+
 // 检查单个平台登录状态
 async function checkPlatformLogin(platform) {
   if (!platform || !platform.id) {
     return { loggedIn: false, error: '无效的平台配置' }
   }
-  const config = LOGIN_CHECK_CONFIG[platform.id]
-  if (!config) {
-    return { loggedIn: false, error: '未配置检测' }
-  }
-
-  // 支付宝开放平台特殊处理：从 storage 读取缓存的用户信息（由 content script 在访问支付宝时缓存）
-  if (platform.id === 'alipayopen') {
-    try {
-      // 从 storage 读取缓存的用户信息
-      const stored = await chrome.storage.local.get('alipayopen_user')
-      const cachedUser = stored.alipayopen_user
-
-      if (cachedUser && cachedUser.loggedIn) {
-        // 检查缓存是否过期（1小时）
-        const cacheAge = Date.now() - (cachedUser.cachedAt || 0)
-        const maxAge = 1 * 60 * 60 * 1000 // 1 hour
-
-        if (cacheAge < maxAge) {
-          console.log(`[COSE] alipayopen 从缓存读取用户信息:`, cachedUser.username)
-          return {
-            loggedIn: true,
-            username: cachedUser.username || '',
-            avatar: cachedUser.avatar || ''
-          }
-        } else {
-          console.log(`[COSE] alipayopen 缓存已过期`)
-          // 清除过期缓存
-          await chrome.storage.local.remove('alipayopen_user')
-        }
-      }
-
-      // 尝试从已打开的支付宝页面获取用户信息并缓存
-      let tabs = await chrome.tabs.query({ url: 'https://open.alipay.com/*' })
-      if (tabs.length === 0) {
-        tabs = await chrome.tabs.query({ url: 'https://*.alipay.com/*' })
-      }
-
-      if (tabs.length > 0) {
-        try {
-          // 在已打开的页面上下文中调用 API
-          const results = await chrome.scripting.executeScript({
-            target: { tabId: tabs[0].id },
-            func: async () => {
-              try {
-                const response = await fetch('https://developerportal.alipay.com/octopus/service.do', {
-                  method: 'POST',
-                  credentials: 'include',
-                  headers: {
-                    'Accept': 'application/json',
-                    'Content-Type': 'application/x-www-form-urlencoded;charset=utf-8',
-                  },
-                  body: 'data=%5B%7B%7D%5D&serviceName=alipay.open.developerops.forum.user.query',
-                })
-                if (!response.ok) return null
-                return await response.json()
-              } catch (e) {
-                return null
-              }
-            }
-          })
-
-          const data = results?.[0]?.result
-          console.log(`[COSE] alipayopen API 数据:`, data)
-
-          if (data?.stat === 'ok' && data?.data?.isLoginUser === 1) {
-            const username = data.data.nickname || ''
-            const avatar = data.data.avatar || ''
-
-            // 缓存用户信息
-            await chrome.storage.local.set({
-              alipayopen_user: {
-                loggedIn: true,
-                username,
-                avatar,
-                cachedAt: Date.now()
-              }
-            })
-
-            console.log(`[COSE] alipayopen 用户信息:`, username, avatar ? '有头像' : '无头像')
-            return { loggedIn: true, username, avatar }
-          }
-        } catch (e) {
-          console.log(`[COSE] alipayopen 从页面获取用户信息失败:`, e.message)
-        }
-      }
-
-      console.log(`[COSE] alipayopen 未检测到登录状态，请先访问支付宝开放平台`)
-      return { loggedIn: false }
-    } catch (e) {
-      console.log(`[COSE] alipayopen 检测失败:`, e.message)
-      return { loggedIn: false, error: e.message }
-    }
-  }
-
-  // 微博特殊处理：通过 cookie 检测登录，通过 fetch HTML 获取用户信息
-  if (platform.id === 'weibo') {
-    try {
-      // 检查 card.weibo.com 的 SUBP cookie
-      const subpCookie = await chrome.cookies.get({
-        url: 'https://card.weibo.com',
-        name: 'SUBP'
-      })
-
-      // 也检查 ALF cookie
-      const alfCookie = await chrome.cookies.get({
-        url: 'https://card.weibo.com',
-        name: 'ALF'
-      })
-
-      console.log(`[COSE] weibo cookies: SUBP=${!!subpCookie}, ALF=${!!alfCookie}`)
-
-      if (!subpCookie && !alfCookie) {
-        console.log(`[COSE] weibo 未找到登录 cookie，未登录`)
-        return { loggedIn: false }
-      }
-
-      // 有 cookie，通过 fetch HTML 获取用户信息
-      let username = ''
-      let avatar = ''
-
-      try {
-        // 获取所有相关 cookies 并手动添加到请求
-        const weiboCookies = await chrome.cookies.getAll({ domain: '.weibo.com' })
-        const cardCookies = await chrome.cookies.getAll({ domain: 'card.weibo.com' })
-        const sinaCookies = await chrome.cookies.getAll({ domain: '.sina.com.cn' })
-        const allCookies = [...weiboCookies, ...cardCookies, ...sinaCookies]
-        const cookieString = allCookies.map(c => `${c.name}=${c.value}`).join('; ')
-
-        console.log(`[COSE] weibo 获取到 ${allCookies.length} 个 cookies`)
-
-        const response = await fetch('https://card.weibo.com/article/v5/editor', {
-          method: 'GET',
-          headers: {
-            'Cookie': cookieString,
-          },
-          credentials: 'include',
-        })
-        const html = await response.text()
-
-        await logToStorage(`[COSE] weibo HTML 长度: ${html.length}`)
-        if (html.length < 1000) {
-          await logToStorage(`[COSE] weibo HTML content snippet:`, html)
-        } else {
-          await logToStorage(`[COSE] weibo HTML content starts with:`, html.substring(0, 500))
-        }
-
-        // 从 HTML 中提取用户名
-        const nickMatch = html.match(/"nick"\s*:\s*"([^"]+)"/)
-        if (nickMatch) {
-          username = nickMatch[1]
-        } else {
-          // 深度查找 nick
-          const altNickMatch = html.match(/\\"nick\\"\s*:\s*\\"([^\\"]+)\\"/)
-          if (altNickMatch) {
-            username = altNickMatch[1]
-            await logToStorage(`[COSE] weibo Found nick via escaped regex: ${username}`)
-          } else {
-            await logToStorage(`[COSE] weibo Failed to find nick in HTML. Snip:`, html.substring(html.indexOf('nick') - 20, html.indexOf('nick') + 100))
-          }
-        }
-
-        // 从 HTML 中提取头像
-        const avatarMatch = html.match(/"avatar_large"\s*:\s*"([^"]+)"/)
-        if (avatarMatch) {
-          // 处理转义的 URL
-          let rawAvatar = avatarMatch[1].replace(/\\/g, '')
-          const avatarUrl = rawAvatar
-          try {
-            await logToStorage(`[COSE] weibo fetching avatar: ${avatarUrl}`)
-            const avatarRes = await fetch(avatarUrl)
-            if (avatarRes.ok) {
-              const blob = await avatarRes.blob()
-              avatar = await new Promise((resolve) => {
-                const reader = new FileReader()
-                reader.onloadend = () => resolve(reader.result)
-                reader.readAsDataURL(blob)
-              })
-              await logToStorage(`[COSE] weibo avatar converted successfully (length: ${avatar.length})`)
-            } else {
-              await logToStorage(`[COSE] weibo avatar fetch failed with status: ${avatarRes.status}`)
-              avatar = avatarUrl
-            }
-          } catch (e) {
-            await logToStorage(`[COSE] weibo avatar fetch failed: ${e.message}`)
-            avatar = avatarUrl
-          }
-        } else {
-          // 深度查找 avatar_large
-          const altAvatarMatch = html.match(/\\"avatar_large\\"\s*:\s*\\"([^\\"]+)\\"/)
-          if (altAvatarMatch) {
-            let rawAvatar = altAvatarMatch[1].replace(/\\\\\\\//g, '/')
-            if (rawAvatar.includes('sinaimg.cn')) {
-              avatar = rawAvatar.split('?')[0]
-            } else {
-              avatar = rawAvatar
-            }
-            await logToStorage(`[COSE] weibo Found avatar via escaped regex: ${avatar}`)
-
-            // Try fetching if it's not converted yet
-            if (avatar && avatar.startsWith('http')) {
-              try {
-                await logToStorage(`[COSE] weibo fetching avatar (alt): ${avatar}`)
-                const avatarRes = await fetch(avatar)
-                if (avatarRes.ok) {
-                  const blob = await avatarRes.blob()
-                  avatar = await new Promise((resolve) => {
-                    const reader = new FileReader()
-                    reader.onloadend = () => resolve(reader.result)
-                    reader.readAsDataURL(blob)
-                  })
-                  await logToStorage(`[COSE] weibo avatar converted successfully (alt) (length: ${avatar.length})`)
-                }
-              } catch (e) {
-                await logToStorage(`[COSE] weibo avatar fetch error (alt): ${e.message}`)
-              }
-            }
-          } else {
-            await logToStorage(`[COSE] weibo Failed to find avatar_large in HTML`)
-          }
-        }
-
-        await logToStorage(`[COSE] weibo 用户信息: ${username} ${avatar ? (avatar.startsWith('data:') ? 'Base64头像' : 'URL头像') : '无头像'}`)
-      } catch (e) {
-        console.log(`[COSE] weibo 获取用户详情失败:`, e.message)
-      }
-
-      // 如果没有获取到用户名，说明实际上未登录或登录已过期
-      if (!username) {
-        console.log(`[COSE] weibo 未获取到用户名，视为未登录`)
-        return { loggedIn: false }
-      }
-
-      return { loggedIn: true, username, avatar }
-    } catch (e) {
-      console.log(`[COSE] weibo 检测失败:`, e.message)
-      return { loggedIn: false }
-    }
-  }
-
-
-
-
-
-
-  console.log(`[COSE] checkPlatformLogin checking: '${platform.id}'`)
-
-  if (platform.id === 'twitter') {
-    console.log(`[COSE] Entering Twitter block for ${platform.id}`)
-    return await checkTwitterLogin(platform)
-  }
-
-  if (config.useCookie) {
-    return await checkLoginByCookie(platform.id, config)
-  }
-
-  // 微信公众号特殊处理：通过已打开的页面或 fetch 首页检测登录状态
-  if (platform.id === 'wechat') {
-    try {
-      // 先检查缓存
-      const stored = await chrome.storage.local.get('wechat_user')
-      const cachedUser = stored.wechat_user
-
-      if (cachedUser && cachedUser.loggedIn) {
-        const cacheAge = Date.now() - (cachedUser.cachedAt || 0)
-        const maxAge = 1 * 60 * 60 * 1000 // 1 hour
-
-        if (cacheAge < maxAge) {
-          console.log(`[COSE] wechat 从缓存读取:`, cachedUser.username)
-          return {
-            loggedIn: true,
-            username: cachedUser.username || '',
-            avatar: cachedUser.avatar || ''
-          }
-        } else {
-          await chrome.storage.local.remove('wechat_user')
-        }
-      }
-
-      // 优先尝试在已打开的微信公众号页面中检测
-      const tabs = await chrome.tabs.query({ url: 'https://mp.weixin.qq.com/*' })
-      if (tabs.length > 0) {
-        try {
-          const results = await chrome.scripting.executeScript({
-            target: { tabId: tabs[0].id },
-            func: () => {
-              // 从 window.wx.data 读取用户信息
-              const wxData = window.wx?.data
-              if (wxData && wxData.nick_name) {
-                return {
-                  loggedIn: true,
-                  username: wxData.nick_name || wxData.user_name || '',
-                  avatar: wxData.head_img || '',
-                  token: wxData.t || ''
-                }
-              }
-              return null
-            }
-          })
-
-          const result = results?.[0]?.result
-          if (result && result.loggedIn) {
-            // 缓存结果
-            const userInfo = {
-              ...result,
-              cachedAt: Date.now()
-            }
-            await chrome.storage.local.set({ wechat_user: userInfo })
-            console.log(`[COSE] wechat 从页面检测成功:`, userInfo.username)
-            return {
-              loggedIn: true,
-              username: userInfo.username || '',
-              avatar: userInfo.avatar || ''
-            }
-          }
-        } catch (e) {
-          console.log(`[COSE] wechat 页面脚本执行失败:`, e.message)
-        }
-      }
-
-      // 备用方案：fetch 首页并解析 HTML
-      try {
-        const response = await fetch('https://mp.weixin.qq.com/', {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            'Accept': 'text/html'
-          }
-        })
-        const html = await response.text()
-
-        // 检查是否需要登录
-        if (html.includes('请使用微信扫描') || html.includes('扫码登录')) {
-          console.log(`[COSE] wechat 未登录（需要扫码）`)
-          return { loggedIn: false }
-        }
-
-        // 尝试从 HTML 提取用户信息
-        const nickMatch = html.match(/nick_name\s*[:=]\s*["']([^"']+)["']/)
-        const avatarMatch = html.match(/head_img\s*[:=]\s*["']([^"']+)["']/)
-
-        if (nickMatch) {
-          const username = nickMatch[1]
-          const avatar = avatarMatch ? avatarMatch[1] : ''
-
-          // 缓存结果
-          await chrome.storage.local.set({
-            wechat_user: {
-              loggedIn: true,
-              username,
-              avatar,
-              cachedAt: Date.now()
-            }
-          })
-
-          console.log(`[COSE] wechat 从 HTML 检测成功:`, username)
-          return { loggedIn: true, username, avatar }
-        }
-      } catch (e) {
-        console.log(`[COSE] wechat fetch 失败:`, e.message)
-      }
-
-      console.log(`[COSE] wechat 未登录或检测失败`)
-      return { loggedIn: false }
-    } catch (e) {
-      console.log(`[COSE] wechat 检测失败:`, e.message)
-      return { loggedIn: false }
-    }
-  }
-
-  // 小红书特殊处理：直接使用 scripting API 检测登录状态
-  if (platform.id === 'xiaohongshu') {
-    try {
-      // 先检查缓存
-      const stored = await chrome.storage.local.get('xiaohongshu_user')
-      const cachedUser = stored.xiaohongshu_user
-
-      if (cachedUser && cachedUser.loggedIn) {
-        const cacheAge = Date.now() - (cachedUser.cachedAt || 0)
-        const maxAge = 7 * 24 * 60 * 60 * 1000 // 7 days
-
-        if (cacheAge < maxAge) {
-          console.log(`[COSE] xiaohongshu 从缓存读取:`, cachedUser.username)
-          return {
-            loggedIn: true,
-            username: cachedUser.username || '',
-            avatar: cachedUser.avatar || ''
-          }
-        } else {
-          await chrome.storage.local.remove('xiaohongshu_user')
-        }
-      }
-
-      // 缓存无效，尝试在已打开的小红书页面中检测
-      const tabs = await chrome.tabs.query({ url: 'https://creator.xiaohongshu.com/*' })
-      if (tabs.length > 0) {
-        const results = await chrome.scripting.executeScript({
-          target: { tabId: tabs[0].id },
-          func: async () => {
-            try {
-              const response = await fetch('https://creator.xiaohongshu.com/api/galaxy/user/info', {
-                method: 'GET',
-                credentials: 'include',
-                headers: { 'Accept': 'application/json' }
-              })
-
-              if (!response.ok) return null
-
-              const data = await response.json()
-              if (data?.success === true && data?.code === 0 && data?.data?.userId) {
-                return {
-                  loggedIn: true,
-                  username: data.data.userName || data.data.redId || '',
-                  avatar: data.data.userAvatar || '',
-                  userId: data.data.userId
-                }
-              }
-              return null
-            } catch (e) {
-              return null
-            }
-          }
-        })
-
-        const result = results?.[0]?.result
-        if (result && result.loggedIn) {
-          // 缓存结果
-          const userInfo = {
-            ...result,
-            cachedAt: Date.now()
-          }
-          await chrome.storage.local.set({ xiaohongshu_user: userInfo })
-          console.log(`[COSE] xiaohongshu 检测成功:`, userInfo.username)
-          return {
-            loggedIn: true,
-            username: userInfo.username || '',
-            avatar: userInfo.avatar || ''
-          }
-        }
-      }
-
-      console.log(`[COSE] xiaohongshu 未登录，请先访问小红书创作者中心`)
-      return { loggedIn: false }
-    } catch (e) {
-      console.log(`[COSE] xiaohongshu 检测失败:`, e.message)
-      return { loggedIn: false }
-    }
-  }
-
-  // 电子发烧友特殊处理：直接调用 API 检测登录状态
-  if (platform.id === 'elecfans') {
-    try {
-      // 先获取 elecfans 的登录 cookie
-      const authCookie = await chrome.cookies.get({
-        url: 'https://www.elecfans.com',
-        name: 'auth'
-      })
-
-      const authWwwCookie = await chrome.cookies.get({
-        url: 'https://www.elecfans.com',
-        name: 'auth_www'
-      })
-
-      console.log(`[COSE] elecfans cookies: auth=${!!authCookie}, auth_www=${!!authWwwCookie}`)
-
-      if (!authCookie && !authWwwCookie) {
-        console.log(`[COSE] elecfans 未找到登录 cookie，未登录`)
-        return { loggedIn: false }
-      }
-
-      // 直接调用 API 获取用户信息
-      try {
-        const response = await fetch('https://www.elecfans.com/webapi/passport/checklogin?_=' + Date.now(), {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            'Accept': 'application/json, text/javascript, */*; q=0.01',
-          }
-        })
-
-        if (!response.ok) {
-          console.log(`[COSE] elecfans API 响应错误:`, response.status)
-          return { loggedIn: true, username: '', avatar: '' }
-        }
-
-        const data = await response.json()
-        console.log(`[COSE] elecfans API 数据:`, data)
-
-        // API 返回格式: {"uid":"6997925","username":"jf_50332692","avatar":"https://..."}
-        if (data && data.uid) {
-          const username = data.username || ''
-          const avatar = data.avatar || ''
-          console.log(`[COSE] elecfans 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          // 有 cookie 但 API 返回无用户，仍然认为已登录
-          console.log(`[COSE] elecfans API 返回无用户数据，但有 cookie`)
-          return { loggedIn: true, username: '', avatar: '' }
-        }
-      } catch (e) {
-        console.log(`[COSE] elecfans API 调用失败:`, e.message)
-        // API 失败但有 cookie，仍然认为已登录
-        return { loggedIn: true, username: '', avatar: '' }
-      }
-    } catch (e) {
-      console.log(`[COSE] elecfans 检测失败:`, e.message)
-      return { loggedIn: false }
-    }
-  }
-
-  // 少数派特殊处理：通过 localStorage 检测登录状态（类似 Sohu）
-  // 少数派特殊处理：通过 Cookie 获取 Token 调用 API
-  if (platform.id === 'sspai') {
-    try {
-      console.log(`[COSE] sspai 开始检测 via API`)
-      const cookie = await chrome.cookies.get({ url: 'https://sspai.com', name: 'sspai_jwt_token' })
-      if (cookie && cookie.value) {
-        const token = cookie.value
-        try {
-          const response = await fetch('https://sspai.com/api/v1/user/info/get', {
-            headers: {
-              'Authorization': `Bearer ${token}`,
-              'Accept': 'application/json'
-            }
-          })
-
-          if (!response.ok) {
-            await logToStorage(`[COSE] sspai API 响应错误: ${response.status}`)
-            return { loggedIn: false }
-          }
-
-          const data = await response.json()
-          await logToStorage(`[COSE] sspai API 数据:`, data)
-
-          if (data && data.data && data.data.nickname) {
-            const userInfo = data.data
-            let avatar = userInfo.avatar || ''
-
-            // 获取头像并转为 base64 解决防盗链
-            if (avatar) {
-              try {
-                await logToStorage(`[COSE] sspai fetching avatar: ${avatar}`)
-                const avatarRes = await fetch(avatar)
-                if (avatarRes.ok) {
-                  const blob = await avatarRes.blob()
-                  avatar = await new Promise((resolve) => {
-                    const reader = new FileReader()
-                    reader.onloadend = () => resolve(reader.result)
-                    reader.readAsDataURL(blob)
-                  })
-                  await logToStorage(`[COSE] sspai avatar converted successfully (length: ${avatar.length})`)
-                }
-              } catch (e) {
-                await logToStorage(`[COSE] sspai avatar fetch failed: ${e.message}`)
-              }
-            }
-
-            await logToStorage(`[COSE] sspai API 检测成功: ${userInfo.nickname}`)
-            return {
-              loggedIn: true,
-              username: userInfo.nickname,
-              avatar: avatar
-            }
-          }
-        } catch (e) {
-          await logToStorage(`[COSE] sspai API fetch error: ${e.message}`)
-        }
-      }
-
-      console.log(`[COSE] sspai API 检测失败 (无有效 Token/Cookie)`)
-      return { loggedIn: false }
-    } catch (e) {
-      console.log(`[COSE] sspai 检测失败:`, e.message)
-      return { loggedIn: false }
-    }
-  }
-
-  // Twitter specialized detection
-  // Twitter specialized detection
-  async function checkTwitterLogin(platform) {
-    try {
-      await logToStorage(`[COSE] Twitter specialized detection started`)
-
-      // 1. Get ct0 cookie
-      const ct0Cookie = await chrome.cookies.get({ url: 'https://x.com', name: 'ct0' })
-      if (!ct0Cookie) {
-        await logToStorage(`[COSE] Twitter detection failed: ct0 cookie not found`)
-        return { loggedIn: false }
-      }
-
-      // 2. Constants
-      const bearerToken = 'Bearer AAAAAAAAAAAAAAAAAAAAANRILgAAAAAAnNwIzUejRCOuH5E6I8xnZz4puTs%3D1Zv7ttfk8LF81IUq16cHjhLTvJu4FA33AGWWjCpTnA'
-
-      let screenName = null
-      let avatar = null
-
-      // 3. Try Login via settings API (Fast Path)
-      try {
-        const settingsResponse = await fetch('https://api.x.com/1.1/account/settings.json', {
-          headers: {
-            'authorization': bearerToken,
-            'x-csrf-token': ct0Cookie.value
-          }
-        })
-
-        if (settingsResponse.ok) {
-          const settingsData = await settingsResponse.json()
-          screenName = settingsData.screen_name
-          await logToStorage(`[COSE] Twitter settings API success: ${screenName}`)
-        } else {
-          await logToStorage(`[COSE] Twitter settings API failed: ${settingsResponse.status}, trying fallback`)
-        }
-      } catch (e) {
-        await logToStorage(`[COSE] Twitter API error: ${e.message}`)
-      }
-
-
-      // 4. Fallback / Avatar Scraping
-      // If we don't have screenName (API failed) or we just need avatar (API doesn't return it easily without more calls)
-      // We explicitly fetch home to scrape.
-
-      if (!screenName || !avatar) {
-        try {
-          const homeResponse = await fetch('https://x.com/home')
-          const homeText = await homeResponse.text()
-
-          // Scrape Avatar
-          const avatarMatch = homeText.match(/"profile_image_url_https":"([^"]+)"/)
-          if (avatarMatch && avatarMatch[1]) {
-            avatar = avatarMatch[1].replace(/\\/g, '')
-          }
-
-          // Scrape Screen Name if missing
-          if (!screenName) {
-            // Try to find screen_name in the initial state
-            // often format: "screen_name":"username"
-            const nameMatch = homeText.match(/"screen_name":"([^"]+)"/)
-            if (nameMatch && nameMatch[1]) {
-              screenName = nameMatch[1]
-            }
-          }
-        } catch (e) {
-          await logToStorage(`[COSE] Twitter scraping failed: ${e.message}`)
-        }
-      }
-
-      if (!screenName) {
-        await logToStorage(`[COSE] Twitter detection failed: screen_name not found via API or Scraping`)
-        return { loggedIn: false }
-      }
-
-      return {
-        loggedIn: true,
-        username: screenName,
-        avatar: avatar
-      }
-
-    } catch (error) {
-      await logToStorage(`[COSE] Twitter detection error: ${error.message}`)
-      return { loggedIn: false, error: error.message }
-    }
-  }
-
-  // Special handling for Twitter
-  if (platform.id === 'twitter') {
-    return await checkTwitterLogin(platform)
-  }
-
-  try {
-    await logToStorage(`[COSE] ${platform.id} 开始 API 检测: ${config.api}`)
-    const controller = new AbortController()
-    const timeoutId = setTimeout(() => controller.abort(), 8000)
-    const response = await fetch(config.api, {
-      method: config.method || 'GET',
-      credentials: 'include',
-      headers: {
-        'Accept': config.isHtml ? 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8' : 'application/json',
-        'Cache-Control': 'no-cache'
-      },
-      signal: controller.signal,
-    })
-    clearTimeout(timeoutId)
-    await logToStorage(`[COSE] ${platform.id} API 响应状态: ${response.status}`)
-
-    let data = null
-    const contentType = response.headers.get('content-type') || ''
-    if (config.isHtml) {
-      // 明确指定为 HTML 响应时，使用 text() 解析
-      try { data = await response.text() } catch (e) { data = '' }
-    } else {
-      // 其他情况尝试 JSON 解析
-      try { data = await response.json() } catch (e) { data = null }
-    }
-    await logToStorage(`[COSE] ${platform.id} API 数据:`, data)
-
-    const loggedIn = config.checkLogin(data)
-    await logToStorage(`[COSE] ${platform.id} checkLogin 结果: ${loggedIn} type: ${typeof loggedIn}`)
-    if (loggedIn && config.getUserInfo) {
-      const userInfo = config.getUserInfo(data)
-      console.log(`[COSE] ${platform.id} 用户信息:`, userInfo)
-      return { loggedIn: true, ...userInfo }
-    }
-    return { loggedIn: !!loggedIn }
-  } catch (error) {
-    await logToStorage(`[COSE] ${platform.id} API 检测失败: ${error.message}`)
-    return { loggedIn: false, error: error.message }
-  }
+  return await detectUser(platform.id)
 }
-
-// 通过 Cookie 检测登录状态
-async function checkLoginByCookie(platformId, config) {
-  try {
-    // 直接按名称查找 cookie
-    const cookieMap = {}
-    for (const name of config.cookieNames) {
-      const cookie = await chrome.cookies.get({
-        url: config.cookieUrl || `https://${config.cookieDomain}`,
-        name: name
-      })
-      if (cookie) {
-        cookieMap[name] = cookie.value
-      }
-    }
-
-    console.log(`[COSE] ${platformId} 找到的cookies:`, Object.keys(cookieMap))
-
-    const hasLoginCookie = config.cookieNames.some(name => cookieMap[name])
-
-    if (!hasLoginCookie) {
-      console.log(`[COSE] ${platformId} 未找到登录 cookie`)
-      return { loggedIn: false }
-    }
-
-    // 自定义 cookie 值检测逻辑（如 InfoQ）
-    if (config.customCheck && config.checkCookieValue) {
-      console.log(`[COSE] ${platformId} 使用自定义 cookie 检测`)
-      const result = config.checkCookieValue(cookieMap)
-      console.log(`[COSE] ${platformId} 自定义检测结果:`, result)
-      return result
-    }
-
-    let username = ''
-    let avatar = ''
-
-    // 如果配置了从 cookie 获取用户名
-    if (config.getUsernameFromCookie && config.usernameCookie) {
-      username = decodeURIComponent(cookieMap[config.usernameCookie] || '')
-    }
-
-
-    // 使用平台配置的 fetchAvatar 回调获取头像
-    if (config.fetchAvatar && typeof config.fetchAvatar === 'function') {
-      try {
-        const fetchedAvatar = await config.fetchAvatar(cookieMap)
-        if (fetchedAvatar) {
-          avatar = fetchedAvatar
-          console.log(`[COSE] ${platformId} 找到头像:`, avatar)
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取头像失败:`, e.message)
-      }
-    }
-
-    // 百家号特殊处理：通过 API 获取用户信息
-    if (platformId === 'baijiahao') {
-      try {
-        // 百家号的 API 需要从页面获取 token，这里直接请求一个不需要 token 的页面
-        // 然后从返回的 HTML 中提取用户信息
-        const response = await fetch('https://baijiahao.baidu.com/builder/app/appinfo', {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            'Accept': 'application/json',
-          }
-        })
-        const data = await response.json()
-
-        if (data.errno === 0 && data.data?.user?.username) {
-          username = data.data.user.username || data.data.user.name
-          avatar = data.data.user.avatar || ''
-          if (avatar.startsWith('//')) {
-            avatar = 'https:' + avatar
-          }
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          console.log(`[COSE] ${platformId} API 返回未登录状态`)
-          return { loggedIn: false }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        // 如果 API 失败，但有 BDUSS cookie，仍然认为已登录
-        return { loggedIn: true, username: '', avatar: '' }
-      }
-    }
-
-    // 网易号特殊处理：通过 navinfo.do API 获取用户信息
-    if (platformId === 'wangyihao') {
-      try {
-        const timestamp = Date.now()
-        const response = await fetch(`https://mp.163.com/wemedia/navinfo.do?_=${timestamp}`, {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            'Accept': 'application/json',
-          }
-        })
-        const data = await response.json()
-
-        if (data.code === 1 && data.data?.tname) {
-          username = data.data.tname
-          avatar = data.data.icon || ''
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          console.log(`[COSE] ${platformId} API 返回未登录状态`)
-          return { loggedIn: false }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        // 如果 API 失败，但有登录 cookie，仍然认为已登录
-        return { loggedIn: true, username: '', avatar: '' }
-      }
-    }
-
-    // 搜狐号特殊处理：通过 API 获取用户信息
-    if (platformId === 'sohu') {
-      try {
-        // 检查 ppinf cookie 判断是否登录
-        const ppinfCookie = await chrome.cookies.get({
-          url: 'https://mp.sohu.com',
-          name: 'ppinf'
-        })
-
-        if (!ppinfCookie || !ppinfCookie.value) {
-          console.log(`[COSE] ${platformId} 未找到 ppinf cookie，未登录`)
-          return { loggedIn: false }
-        }
-
-        // 直接调用 API 获取用户信息
-        try {
-          const response = await fetch('https://mp.sohu.com/mpbp/bp/account/list', {
-            method: 'GET',
-            credentials: 'include',
-            headers: {
-              'Accept': 'application/json',
-            }
-          })
-          const data = await response.json()
-
-          if (data.success && data.data?.data?.[0]?.accounts?.[0]) {
-            const account = data.data.data[0].accounts[0]
-            let avatar = account.avatar || ''
-            if (avatar.startsWith('//')) {
-              avatar = 'https:' + avatar
-            }
-            console.log(`[COSE] ${platformId} 用户信息:`, account.nickName, avatar ? '有头像' : '无头像')
-            return {
-              loggedIn: true,
-              username: account.nickName,
-              avatar
-            }
-          } else {
-            console.log(`[COSE] ${platformId} API 返回无用户数据`)
-            return { loggedIn: true, username: '', avatar: '' }
-          }
-        } catch (e) {
-          console.log(`[COSE] ${platformId} API 调用失败:`, e.message)
-          // API 失败但有 cookie，仍然认为已登录
-          return { loggedIn: true, username: '', avatar: '' }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // 阿里云开发者社区特殊处理：通过 API 获取用户信息
-    if (platformId === 'aliyun') {
-      try {
-        // 检查 login_aliyunid_ticket cookie 判断是否登录
-        const ticketCookie = await chrome.cookies.get({
-          url: 'https://developer.aliyun.com',
-          name: 'login_aliyunid_ticket'
-        })
-
-        if (!ticketCookie || !ticketCookie.value) {
-          console.log(`[COSE] ${platformId} 未找到 login_aliyunid_ticket cookie，未登录`)
-          return { loggedIn: false }
-        }
-
-        // 调用 API 获取用户信息
-        const response = await fetch('https://developer.aliyun.com/developer/api/my/user/getUser', {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            'Accept': 'application/json',
-          }
-        })
-        const data = await response.json()
-
-        if (data.success && data.data?.nickname) {
-          const username = data.data.nickname
-          const avatar = data.data.avatar || ''
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          console.log(`[COSE] ${platformId} API 返回未登录状态`)
-          return { loggedIn: false }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // 华为云开发者博客特殊处理
-    if (platformId === 'huaweicloud') {
-      try {
-        // 查找华为云相关页面（优先 bbs，其次其他华为云页面）
-        let tabs = await chrome.tabs.query({ url: 'https://bbs.huaweicloud.com/*' })
-
-        if (tabs.length === 0) {
-          tabs = await chrome.tabs.query({ url: 'https://*.huaweicloud.com/*' })
-        }
-
-        if (tabs.length === 0) {
-          console.log(`[COSE] ${platformId} 没有打开的华为云页面`)
-          return { loggedIn: false }
-        }
-
-        console.log(`[COSE] ${platformId} 使用页面:`, tabs[0].url)
-
-        // 在页面上下文中调用 API（使用同步包装）
-        try {
-          const results = await chrome.scripting.executeScript({
-            target: { tabId: tabs[0].id },
-            func: () => {
-              return new Promise((resolve) => {
-                const csrf = document.cookie.match(/csrf=([^;]+)/)?.[1] || ''
-                fetch('https://devdata.huaweicloud.com/rest/developer/fwdu/rest/developer/user/hdcommunityservice/v1/member/get-personal-info', {
-                  method: 'GET',
-                  credentials: 'include',
-                  headers: { 'Accept': 'application/json', 'csrf': csrf }
-                })
-                  .then(response => response.ok ? response.json() : null)
-                  .then(data => {
-                    if (data && data.memName) {
-                      resolve({ memName: data.memName, memAlias: data.memAlias, memPhoto: data.memPhoto })
-                    } else {
-                      resolve(null)
-                    }
-                  })
-                  .catch(() => resolve(null))
-              })
-            }
-          })
-
-          // executeScript 返回的 result 可能是 Promise，需要 await
-          let data = results?.[0]?.result
-          if (data && typeof data.then === 'function') {
-            data = await data
-          }
-
-          if (data && data.memName) {
-            console.log(`[COSE] ${platformId} 已登录:`, data.memName)
-            return {
-              loggedIn: true,
-              username: data.memAlias || data.memName,
-              avatar: data.memPhoto || ''
-            }
-          }
-          console.log(`[COSE] ${platformId} API 返回:`, data)
-        } catch (e) {
-          console.log(`[COSE] ${platformId} scripting 失败:`, e.message)
-        }
-
-        return { loggedIn: false }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 检测失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // 华为开发者文章特殊处理
-    if (platformId === 'huaweidev') {
-      try {
-        // 检查 developer_userinfo cookie
-        const userInfoCookie = await chrome.cookies.get({
-          url: 'https://developer.huawei.com',
-          name: 'developer_userinfo'
-        })
-
-        if (!userInfoCookie || !userInfoCookie.value) {
-          console.log(`[COSE] ${platformId} 未找到 developer_userinfo cookie，未登录`)
-          return { loggedIn: false }
-        }
-
-        console.log(`[COSE] ${platformId} 找到 developer_userinfo cookie`)
-
-        // 解析 cookie 获取 csrfToken
-        let csrfToken = ''
-        try {
-          const userInfoData = JSON.parse(decodeURIComponent(userInfoCookie.value))
-          csrfToken = userInfoData.csrftoken || ''
-        } catch (e) {
-          console.log(`[COSE] ${platformId} 解析 cookie 失败:`, e.message)
-        }
-
-        // 如果没有 csrfToken，尝试从单独的 cookie 获取
-        if (!csrfToken) {
-          const csrfCookie = await chrome.cookies.get({
-            url: 'https://developer.huawei.com',
-            name: 'csrfToken'
-          })
-          csrfToken = csrfCookie?.value || ''
-        }
-
-        if (!csrfToken) {
-          console.log(`[COSE] ${platformId} 未找到 csrfToken，返回已登录但无用户信息`)
-          return { loggedIn: true, username: '', avatar: '' }
-        }
-
-        // 通过 API 获取用户信息
-        const now = new Date()
-        const hdDate = now.toISOString().replace(/[-:]/g, '').replace(/\.\d{3}/, '')
-        const serialNo = Math.floor(Math.random() * 10000000).toString()
-
-        const response = await fetch('https://svc-drcn.developer.huawei.com/codeserver/Common/v1/delegate', {
-          method: 'POST',
-          credentials: 'include',
-          headers: {
-            'Accept': 'application/json, text/plain, */*',
-            'Content-Type': 'application/json;charset=UTF-8',
-            'x-hd-csrf': csrfToken,
-            'x-hd-date': hdDate,
-            'x-hd-serialno': serialNo,
-            'Origin': 'https://developer.huawei.com',
-            'Referer': 'https://developer.huawei.com/'
-          },
-          body: JSON.stringify({
-            svc: 'GOpen.User.getInfo',
-            reqType: 0,
-            reqJson: JSON.stringify({ getNickName: '1' })
-          })
-        })
-
-        const data = await response.json()
-
-        if (data.returnCode === '0' && data.resJson) {
-          const userInfo = JSON.parse(data.resJson)
-          const username = userInfo.displayName || userInfo.loginID || ''
-          const avatar = userInfo.headPictureURL || ''
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          console.log(`[COSE] ${platformId} API 返回错误:`, data.returnCode, data.description)
-          return { loggedIn: true, username: '', avatar: '' }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 检测失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // 少数派特殊处理：通过 /api/v1/user/info/get API 获取用户信息
-    // 需要从 cookie 中读取 JWT token 并添加到 Authorization header
-    if (platformId === 'sspai') {
-      try {
-        // 先获取 JWT token
-        const jwtCookie = await chrome.cookies.get({
-          url: 'https://sspai.com',
-          name: 'sspai_jwt_token'
-        })
-
-        if (!jwtCookie || !jwtCookie.value) {
-          console.log(`[COSE] ${platformId} 未找到 JWT token`)
-          return { loggedIn: false }
-        }
-
-        const token = jwtCookie.value
-        const response = await fetch('https://sspai.com/api/v1/user/info/get', {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            'Accept': 'application/json',
-            'Authorization': `Bearer ${token}`,
-          }
-        })
-        const data = await response.json()
-
-        if (data.error === 0 && data.data?.nickname) {
-          username = data.data.nickname
-          // 少数派 CDN 没有跨域限制，直接使用原始头像 URL
-          avatar = data.data.avatar || ''
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          console.log(`[COSE] ${platformId} API 返回未登录状态:`, data.msg || data.error)
-          return { loggedIn: false }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // 腾讯云开发者社区特殊处理：通过创作中心页面获取用户信息
-    if (platformId === 'tencentcloud') {
-      try {
-        const response = await fetch('https://cloud.tencent.com/developer/creator', {
-          method: 'GET',
-          credentials: 'include',
-        })
-        const html = await response.text()
-        const finalUrl = response.url
-
-        // 检查是否被重定向到首页（未登录时会重定向）
-        if (!finalUrl.includes('/creator')) {
-          console.log(`[COSE] ${platformId} 未登录：被重定向到 ${finalUrl}`)
-          return { loggedIn: false }
-        }
-
-        // 检查页面是否包含登录按钮（未登录标志）
-        if (html.includes('登录/注册') || html.includes('"isLogin":false') || html.includes('"login":false')) {
-          console.log(`[COSE] ${platformId} 未登录：页面包含登录按钮`)
-          return { loggedIn: false }
-        }
-
-        // 从创作中心页面提取当前用户信息（通常在页面的用户信息区域）
-        // 匹配 "userInfo" 或 "creatorInfo" 等包含当前用户信息的 JSON 对象
-        const userInfoMatch = html.match(/"userInfo"\s*:\s*\{[^}]*"nickname"\s*:\s*"([^"]+)"[^}]*\}/) ||
-          html.match(/"creatorInfo"\s*:\s*\{[^}]*"nickname"\s*:\s*"([^"]+)"[^}]*\}/) ||
-          html.match(/"currentUser"\s*:\s*\{[^}]*"nickname"\s*:\s*"([^"]+)"[^}]*\}/)
-
-        // 备用方案：匹配创作中心特有的用户信息结构
-        const creatorNicknameMatch = html.match(/class="creator-info[^"]*"[^>]*>[\s\S]*?<[^>]*class="[^"]*name[^"]*"[^>]*>([^<]+)</) ||
-          html.match(/"isCreator"\s*:\s*true[\s\S]*?"nickname"\s*:\s*"([^"]+)"/)
-
-        const nicknameMatch = userInfoMatch || creatorNicknameMatch
-        const avatarMatch = html.match(/"userInfo"[\s\S]*?"avatarUrl"\s*:\s*"([^"]+)"/) ||
-          html.match(/"avatar"\s*:\s*"(https?:\/\/[^"]+)"/)
-
-        if (nicknameMatch && nicknameMatch[1]) {
-          username = nicknameMatch[1]
-          avatar = avatarMatch ? avatarMatch[1] : ''
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          // 如果在创作中心页面但无法提取用户信息，可能是页面结构变化
-          // 检查是否有创作中心特有的元素来确认登录状态
-          if (html.includes('创作中心') || html.includes('我的文章') || html.includes('发布文章')) {
-            console.log(`[COSE] ${platformId} 已登录但无法提取用户名`)
-            return { loggedIn: true, username: '', avatar: '' }
-          }
-          console.log(`[COSE] ${platformId} 未检测到登录状态`)
-          return { loggedIn: false }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // Medium 特殊处理：先检查 cookies，再获取用户信息
-    if (platformId === 'medium') {
-      try {
-        // Medium 使用 sid 和 uid cookies 来标识登录状态
-        const sidCookie = await chrome.cookies.get({
-          url: 'https://medium.com',
-          name: 'sid'
-        })
-        const uidCookie = await chrome.cookies.get({
-          url: 'https://medium.com',
-          name: 'uid'
-        })
-
-        // 如果没有 sid 或 uid cookie，说明未登录
-        if (!sidCookie && !uidCookie) {
-          console.log(`[COSE] ${platformId} 未找到登录 cookie`)
-          return { loggedIn: false }
-        }
-
-        console.log(`[COSE] ${platformId} 找到登录 cookie: sid=${!!sidCookie}, uid=${!!uidCookie}`)
-
-        // 尝试获取用户信息
-        const response = await fetch('https://medium.com/me/stats', {
-          method: 'GET',
-          credentials: 'include',
-        })
-        const html = await response.text()
-        const finalUrl = response.url
-
-        // 检查是否被重定向到登录页（未登录标志）
-        if (finalUrl.includes('/m/signin') || finalUrl.includes('?signIn') || finalUrl.includes('/m/callback')) {
-          console.log(`[COSE] ${platformId} 未登录：被重定向到登录页`)
-          return { loggedIn: false }
-        }
-
-        // 从页面提取用户名 - 优先使用 JSON 格式
-        const profileMatch = html.match(/"username"\s*:\s*"([^"]+)"/) ||
-          html.match(/href="https:\/\/medium\.com\/@([^"?\/]+)"/) ||
-          html.match(/medium\.com\/@([a-zA-Z0-9_]+)/)
-
-        if (profileMatch && profileMatch[1] && profileMatch[1] !== 'gmail' && profileMatch[1] !== 'medium') {
-          username = profileMatch[1]
-
-          // 从页面提取头像 - 查找导航栏用户头像
-          // Medium 的用户头像 img 标签的 alt 属性是用户名的简称（如 "fwen"）
-          // 格式: <img alt="fwen" src="https://miro.medium.com/v2/resize:fill:64:64/...">
-          const shortName = username.replace(/\d+$/, '') // 移除末尾数字，如 fwen925 -> fwen
-          const avatarPattern = new RegExp(`<img[^>]*alt="${shortName}"[^>]*src="([^"]+)"`)
-          const avatarMatch = html.match(avatarPattern)
-
-          // 备用: src 在 alt 之前
-          const avatarPattern2 = new RegExp(`<img[^>]*src="([^"]+)"[^>]*alt="${shortName}"`)
-          const avatarMatch2 = html.match(avatarPattern2)
-
-          avatar = (avatarMatch && avatarMatch[1]) || (avatarMatch2 && avatarMatch2[1]) || ''
-
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          return { loggedIn: true, username, avatar }
-        } else {
-          // 有 cookie 但无法提取用户名，仍然认为已登录
-          console.log(`[COSE] ${platformId} 已登录但无法提取用户名`)
-          return { loggedIn: true, username: '', avatar: '' }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // Twitter/X 特殊处理：检查 auth_token 和 ct0 cookies
-    if (platformId === 'twitter') {
-      try {
-        // Twitter 使用 auth_token 和 ct0 cookies 来标识登录状态
-        const authTokenCookie = await chrome.cookies.get({
-          url: 'https://x.com',
-          name: 'auth_token'
-        })
-        const ct0Cookie = await chrome.cookies.get({
-          url: 'https://x.com',
-          name: 'ct0'
-        })
-
-        // 如果没有 auth_token cookie，说明未登录
-        if (!authTokenCookie) {
-          console.log(`[COSE] ${platformId} 未找到 auth_token cookie，未登录`)
-          return { loggedIn: false }
-        }
-
-        console.log(`[COSE] ${platformId} 找到登录 cookie: auth_token=${!!authTokenCookie}, ct0=${!!ct0Cookie}`)
-
-        // 尝试获取用户信息
-        let username = ''
-        let avatar = ''
-
-        try {
-          // 方法1: 从 Twitter 首页 HTML 中提取用户信息
-          const response = await fetch('https://x.com/home', {
-            method: 'GET',
-            credentials: 'include',
-            headers: {
-              'Accept': 'text/html',
-            }
-          })
-
-          if (response.ok) {
-            const html = await response.text()
-
-            // 从页面脚本中提取 screen_name
-            const screenNameMatch = html.match(/"screen_name"\s*:\s*"([^"]+)"/)
-            if (screenNameMatch) {
-              username = screenNameMatch[1]
-            }
-
-            // 从页面脚本中提取头像 URL
-            const avatarMatch = html.match(/"profile_image_url_https"\s*:\s*"([^"]+)"/)
-            if (avatarMatch) {
-              // 将 _normal 替换为 _x96 获取更大的头像
-              avatar = avatarMatch[1].replace('_normal.', '_x96.')
-            }
-
-            console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-          }
-        } catch (e) {
-          console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-        }
-
-        // 有 auth_token cookie 就认为已登录
-        return { loggedIn: true, username, avatar }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 检测失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // 百度云千帆特殊处理：通过 API 获取用户信息
-    if (platformId === 'qianfan') {
-      try {
-        // 调用千帆社区用户信息 API
-        const response = await fetch('https://qianfan.cloud.baidu.com/api/community/user/current', {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            'Accept': 'application/json',
-          }
-        })
-
-        if (!response.ok) {
-          console.log(`[COSE] ${platformId} API 请求失败: ${response.status}`)
-          return { loggedIn: false }
-        }
-
-        const data = await response.json()
-        console.log(`[COSE] ${platformId} API 响应:`, data)
-
-        // 检查 API 返回是否成功且有用户信息
-        if (data.success && data.result) {
-          const username = data.result.displayName || data.result.nickname || ''
-          const avatar = data.result.avatar || ''
-
-          console.log(`[COSE] ${platformId} 用户信息: ${username}, ${avatar ? '有头像' : '无头像'}`)
-          return { loggedIn: true, username, avatar }
-        } else {
-          console.log(`[COSE] ${platformId} 未登录或无用户信息`)
-          return { loggedIn: false }
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 检测失败:`, e.message)
-        return { loggedIn: false }
-      }
-    }
-
-    // 从页面抓取用户信息
-    if (config.fetchUserInfoFromPage && config.userInfoUrl) {
-      try {
-        const response = await fetch(config.userInfoUrl, {
-          method: 'GET',
-          credentials: 'include'
-        })
-        const html = await response.text()
-
-        // 头条号用户信息提取
-        if (platformId === 'toutiao') {
-          // 尝试从页面中提取用户名
-          const nameMatch = html.match(/\"name\"\s*:\s*\"([^"]+)\"/i) ||
-            html.match(/screen_name[\"']?\s*[:=]\s*[\"']([^\"']+)[\"']/i) ||
-            html.match(/<span[^>]*class="[^"]*name[^"]*"[^>]*>([^<]+)<\/span>/i)
-          if (nameMatch) {
-            username = nameMatch[1]
-          }
-          // 尝试从页面中提取头像
-          const avatarMatch = html.match(/\"avatar_url\"\s*:\s*\"([^"]+)\"/i) ||
-            html.match(/avatar[\"']?\s*[:=]\s*[\"']([^\"']+)[\"']/i)
-          if (avatarMatch) {
-            avatar = avatarMatch[1].replace(/\\/g, '')
-          }
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-        }
-        // 思否用户信息提取（从 __NEXT_DATA__ 中获取）
-        // 注意：思否的 PHPSESSID 即使未登录也存在，必须通过用户信息判断登录状态
-        else if (platformId === 'segmentfault') {
-          // 从 __NEXT_DATA__ 中提取用户信息
-          const nextDataMatch = html.match(/<script\s+id="__NEXT_DATA__"[^>]*>([^<]+)<\/script>/i)
-          if (nextDataMatch) {
-            try {
-              const nextData = JSON.parse(nextDataMatch[1])
-              const sessionUser = nextData?.props?.pageProps?.initialState?.global?.sessionUser?.user
-              if (sessionUser && sessionUser.name) {
-                username = sessionUser.name
-                avatar = sessionUser.avatar_url || ''
-              }
-            } catch (e) {
-              console.log(`[COSE] ${platformId} 解析 NEXT_DATA 失败:`, e.message)
-            }
-          }
-          // 思否：只有成功提取到用户名才算已登录
-          if (!username) {
-            console.log(`[COSE] ${platformId} 未登录或无法获取用户信息`)
-            return { loggedIn: false }
-          }
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-        }
-        // 微信公众号用户信息提取
-        else if (platformId === 'wechat') {
-          // 从 HTML 中提取公众号名称
-          const nameMatch = html.match(/nick_name\s*[:=]\s*["']([^"']+)["']/i) ||
-            html.match(/<span[^>]*class="nickname"[^>]*>([^<]+)<\/span>/i)
-          if (nameMatch) {
-            username = nameMatch[1]
-          }
-          // 从 HTML 中提取头像
-          const avatarMatch = html.match(/head_img\s*[:=]\s*["']([^"']+)["']/i) ||
-            html.match(/<img[^>]*class="avatar"[^>]*src="([^"]+)"/i)
-          if (avatarMatch) {
-            avatar = avatarMatch[1].replace(/\\x26amp;/g, '&').replace(/\\/g, '')
-            if (!avatar.startsWith('http')) {
-              avatar = 'https://mp.weixin.qq.com' + avatar
-            }
-          }
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-        }
-        // OSChina 用户信息提取
-        else if (platformId === 'oschina') {
-          // 提取用户 ID（从个人空间链接）
-          const uidMatch = html.match(/href=["']https:\/\/my\.oschina\.net\/u\/(\d+)["']/i) ||
-            html.match(/space\.oschina\.net\/u\/(\d+)/i) ||
-            html.match(/data-user-id=["'](\d+)["']/i)
-
-          let userId = null
-          if (uidMatch) {
-            userId = uidMatch[1]
-            PLATFORM_USER_INFO['oschina'] = { userId }
-            console.log(`[COSE] ${platformId} 获取到 userId:`, userId)
-          }
-
-          // 从 HTML 中提取用户名
-          const nameMatch = html.match(/<a[^>]*class="[^"]*user-name[^"]*"[^>]*>([^<]+)<\/a>/i) ||
-            html.match(/<span[^>]*class="[^"]*nick[^"]*"[^>]*>([^<]+)<\/span>/i) ||
-            html.match(/title="([^"]+)"[^>]*class="[^"]*avatar/i) ||
-            html.match(/class="[^"]*avatar[^"]*"[^>]*title="([^"]+)"/i) ||
-            html.match(/alt="([^"]+)"[^>]*class="[^"]*avatar/i)
-
-          if (nameMatch) {
-            username = nameMatch[1].trim()
-          }
-
-          // 如果首页没提取到用户名，但有 userId，尝试从个人空间获取
-          if (!username && userId) {
-            try {
-              console.log(`[COSE] ${platformId} 尝试从个人空间获取用户名...`)
-              const userPageResp = await fetch(`https://my.oschina.net/u/${userId}`, {
-                method: 'GET',
-                credentials: 'include'
-              })
-              const userPageHtml = await userPageResp.text()
-              const userPageNameMatch = userPageHtml.match(/<h3[^>]*class="[^"]*header[^"]*"[^>]*>([^<]+)<\/h3>/i) ||
-                userPageHtml.match(/<div[^>]*class="[^"]*user-name[^"]*"[^>]*>([^<]+)<\/div>/i) ||
-                userPageHtml.match(/<title>([^<]+)的个人空间<\/title>/i)
-              if (userPageNameMatch) {
-                username = userPageNameMatch[1].trim()
-                console.log(`[COSE] ${platformId} 从个人空间获取用户名成功:`, username)
-              }
-            } catch (err) {
-              console.error(`[COSE] ${platformId} 获取个人空间失败:`, err)
-            }
-          }
-
-          // 如果还是没有用户名，但有 userId，使用 userId 作为兜底
-          if (!username && userId) {
-            username = userId
-          }
-
-          // 头像提取
-          const avatarMatch = html.match(/<img[^>]*src="([^"]+)"[^>]*class="[^"]*avatar/i) ||
-            html.match(/<img[^>]*class="[^"]*avatar[^"]*"[^>]*src="([^"]+)"/i)
-          if (avatarMatch) {
-            avatar = avatarMatch[1]
-          }
-
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-        }
-        // 通用模块化解析逻辑 (支持 51CTO 等新平台)
-        else if (config.parseUserInfo) {
-          console.log(`[COSE] ${platformId} 使用通用解析逻辑`)
-          const info = config.parseUserInfo(html)
-          if (info) {
-            // 允许解析函数显式返回未登录状态
-            if (info.loggedIn === false) {
-              console.log(`[COSE] ${platformId} 解析结果判定为未登录`)
-              return { loggedIn: false }
-            }
-            if (info.username) username = info.username
-            if (info.avatar) avatar = info.avatar
-            // 如果解析出了 userId，保存到全局存储
-            if (info.userId && typeof PLATFORM_USER_INFO !== 'undefined') {
-              PLATFORM_USER_INFO[platformId] = { userId: info.userId }
-              console.log(`[COSE] ${platformId} 获取到 userId:`, info.userId)
-            }
-          }
-          console.log(`[COSE] ${platformId} 用户信息:`, username, avatar ? '有头像' : '无头像')
-        }
-      } catch (e) {
-        console.log(`[COSE] ${platformId} 获取用户信息失败:`, e.message)
-      }
-    }
-
-    console.log(`[COSE] ${platformId} 登录用户:`, username, avatar ? '有头像' : '无头像')
-    return { loggedIn: true, username, avatar }
-  } catch (error) {
-    console.error(`[COSE] ${platformId} cookie检测错误:`, error)
-    return { loggedIn: false, error: error.message }
-  }
-}
-
-// 使用 Debugger API 发送真实的 Ctrl+V 粘贴
 async function pasteWithDebugger(tabId) {
   const debuggee = { tabId }
